requirements.txt
transformers==4.40.0
datasets==2.18.0
torch==2.2.2
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.2
evaluate==0.4.1
wandb==0.16.6
pyarrow==15.0.2
tensorflow==2.18.0
huggingface-hub==0.22.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
email-validator==2.1.0.post1
beautifulsoup4
PyJWT
